+++ CHANGELOG.md
@@ -1,20 +1,34 @@
 # Changelog
 
+## [1.0.3] - 2025-10-14
+
+### Fixed
+
+- EditorStartInitSettingSO 생성 시 상위 폴더 미존재로 인한 `Parent directory must exist` 오류 해결
+- Custom Debug 더블 클릭 시 로그 호출 위치로 이동하지 않던 문제 해결
+
 ## [1.0.2] - 2025-10-13
+
 ### Added
+
 - ReadOnly Attribute 추가 (인스펙터에서 읽기 전용 필드 표시)
 
 ### Changed
+
 - namespace 변경 Utility -> Toolkit
 - BasicTemplateToolbarExtender -> PackageExtender로 이름 변경
 - AudioExtensions summary 수정
 
 ## [1.0.1] - 2025-10-13
+
 ### Added
+
 - Actions를 이용하여 자동 Push 설정
 
 ### Fixed
+
 - BasicTemplateToolbarExtender 스크립트에서 AddPackage함수 매개변수를 이용하여 패키지 추가하도록 수정
 
 ## [1.0.0] - 2025-10-13
+
 패키지 추가